version 2.1.9: add VS2017 compatiblity to the old vsixmanifest
This version of the sources can be built with VS2005, VS2008 and VS2010.
The extension built with VS2010 is compatible version from VS2010 to VS2017
but VS2017 pops up a compatibility varning message.

To avoid the warning the extension has to be built with the new manifest format
which makes the extension compatible only with VS2012 and newer but this branch
doesn't support that build.